upload results subtask 2

+++ Models/LM/src/utils.py
@@ -23,6 +23,18 @@ def get_config(location):
         print(error)
 
 
+# def read_data(file):
+#     """Read in data sets and returns sentences and labels"""
+#     sentences = []
+#     labels = []
+#     with open(file, encoding='utf-8') as f:
+#         for line in f:
+#             tokens = line.strip().split()
+#             sentences.append(" ".join(tokens[:-1]))
+#             labels.append(int(tokens[-1]))
+#     return sentences, labels
+
+
 def read_data(file):
     """Read in data sets and returns sentences and labels"""
     sentences = []
@@ -31,7 +43,7 @@ def read_data(file):
         for line in f:
             tokens = line.strip().split()
             sentences.append(" ".join(tokens[:-1]))
-            labels.append(int(tokens[-1]))
+            labels.append(float(tokens[-1]))
     return sentences, labels
